+++ cmd/serper/main_test.go
@@ -4,13 +4,13 @@ import (
 	"os"
 	"testing"
 
-	chassis "github.com/ai8future/chassis-go/v8"
-	chassisconfig "github.com/ai8future/chassis-go/v8/config"
-	"github.com/ai8future/chassis-go/v8/testkit"
+	chassis "github.com/ai8future/chassis-go/v9"
+	chassisconfig "github.com/ai8future/chassis-go/v9/config"
+	"github.com/ai8future/chassis-go/v9/testkit"
 )
 
 func TestMain(m *testing.M) {
-	chassis.RequireMajor(8)
+	chassis.RequireMajor(9)
 	code := m.Run()
 	chassis.ResetVersionCheck()
 	os.Exit(code)